Reformat zkproofs module
